+++ CryptocurrencyExchange/Models/Future.cs
@@ -4,7 +4,7 @@
     {
         public int Id { get; set; }
         public string Symbol { get; set; } = string.Empty;
-        public decimal Margin { get; set; }
+        public double Margin { get; set; }
         public double EntryPrice { get; set; }
         public int UserId { get; set; }
         public bool IsCompleted { get; set; }
